requirements.txt
discord.py==2.3.2
aiohttp==3.9.5
topgg==1.6.0

